backend/huggingface_backend/Rag_Chatbot_Deployment/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "humanoid-robotics-rag-agent"
version = "0.1.0"
description = "A RAG agent for humanoid robotics documentation"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "qdrant-client>=1.7.0",
    "cohere>=5.5.3",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "agents>=0.1.0",
    "openai>=1.3.5",
    "trafilatura>=1.6.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0"
]

[tool.setuptools.packages.find]
where = ["."]